feat: Add reconciliation script for quick-flip positions and market preferences utility
- Implemented `reconcile_quick_flip_positions.py` to clean up stale quick-flip positions in the database.
- Created `market_preferences.py` for market category normalization and live-wagering identification.
- Added tests for CLI safety flags, market preferences, quick flip scalping strategy, safe compounder, and trade job configurations.
- Enhanced quick flip scalping strategy with new methods for evaluating price opportunities and managing live positions.
- Improved handling of live trading configurations and ensured proper behavior in paper mode.

+++ src/utils/stop_loss_calculator.py
@@ -172,7 +172,7 @@ class StopLossCalculator:
         cls,
         entry_price: float,
         stop_loss_price: float,
-        quantity: int,
+        quantity: float,
         side: str
     ) -> float:
         """
